fix: add custom hook for activation param Description: Add custom hook for account activation param detection VAN-1945
