+++ src/data/constants.js
@@ -23,5 +23,5 @@ export const ENTERPRISE_LOGIN_URL = '/enterprise/login';
 // Query string parameters that can be passed to LMS to manage
 // things like auto-enrollment upon login and registration.
 export const VALID_AUTH_PARAMS = [
-  'course_id', 'enrollment_action', 'course_mode', 'email_opt_in', 'purchase_workflow', 'next', 'tpa_hint',
+  'course_id', 'enrollment_action', 'course_mode', 'email_opt_in', 'purchase_workflow', 'next', 'tpa_hint', 'account_activation_status',
 ];
