+++ src/forms/login-popup/index.jsx
@@ -10,6 +10,7 @@ import {
 import AccountActivationMessage from './components/AccountActivationMessage';
 import LoginFailureAlert from './components/LoginFailureAlert';
 import { TPA_AUTHENTICATION_FAILURE } from './data/constants';
+import useGetActivationMessage from './data/hooks';
 import { loginUser } from './data/reducers';
 import messages from './messages';
 import { setCurrentOpenedForm } from '../../authn-component/data/reducers';
@@ -42,11 +43,13 @@ const LoginForm = () => {
   const thirdPartyAuthErrorMessage = useSelector(state => state.commonData.thirdPartyAuthContext.errorMessage);
   const finishAuthUrl = useSelector(state => state.commonData.thirdPartyAuthContext.finishAuthUrl);
 
+  const accountActivation = useGetActivationMessage();
+
   const [formFields, setFormFields] = useState({
     emailOrUsername: '',
     password: '',
   });
-  const [showActivationMessage, setActivationMessage] = useState(null);
+
   const [formErrors, setFormErrors] = useState({
     emailOrUsername: '',
     password: '',
@@ -73,17 +76,6 @@ const LoginForm = () => {
       }));
     }
   }, [thirdPartyAuthErrorMessage]);
-
-  useEffect(() => {
-    const url = new URL(window.location.href);
-
-    const accountActivationParam = url.searchParams.get('account_activation_status');
-    if (accountActivationParam) {
-      setActivationMessage(accountActivationParam);
-      url.searchParams.delete('account_activation_status');
-      window.history.replaceState(window.history.state, '', url.href);
-    }
-  }, []);
 
   const validateFormFields = (payload) => {
     const { emailOrUsername, password } = payload;
@@ -156,7 +148,7 @@ const LoginForm = () => {
         currentProvider={currentProvider}
       />
       <AccountActivationMessage
-        messageType={showActivationMessage}
+        messageType={accountActivation}
       />
       <Form id="login-form" name="login-form">
         <EmailOrUsernameField
