+++ src/forms/login-popup/tests/LoginPopup.test.jsx
@@ -352,15 +352,4 @@ describe('LoginForm Test', () => {
 
     expect(store.dispatch).toHaveBeenCalledWith(setCurrentOpenedForm(FORGOT_PASSWORD_FORM));
   });
-
-  it('should remove the account activation param from url', () => {
-    const href = 'localhost:2999/login';
-    delete window.location;
-    window.history.replaceState = jest.fn();
-    window.location = { href: `${href}?account_activation_status=success` };
-
-    render(reduxWrapper(<IntlLoginForm />));
-
-    expect(window.history.replaceState).toHaveBeenCalledTimes(1);
-  });
 });
